panda: Fix SerialPollGetControl with returning correct Control parm
This function has been added recently as an empty stub but actually
causing an issue because Control parameter is not updated and function
returns with no error.

The side effect is that with DEBUG release the Control parameter is
returned with junk data making the console no longer accepting
characters. Interestingly it did not appear in RELEASE target
probably because Control gets updated with some other random
data making the upper level function react correctly.