ArmPlatformPkg/Bds: add local FDT config
This patch adds a local FDT config option to each boot device.

Bds was using a global FDT config for all boot devices of Loader Type = 2.

I've added a new Loader Type = 3 that stores an device path to an FDT file this is only used for this boot device.

This change will invalidate any existing configs stored on the board due to the change in structure sizes to store the boot device configurations.

Signed-off-by: Ryan Harkin <ryan.harkin@linaro.org>
Signed-off-by: Steven Kinney <steven.kinney@linaro.org>