ArmPlatformPkg/Bds: add local FDT config
This patch adds a local FDT config option to each boot device.

Bds was using a global FDT config for all boot devices of Loader Type = 2.

I've added a new Loader Type = 3 that stores an device path to an FDT file this is only used for this boot device.

This change will invalidate any existing configs stored on the board due to the change in structure sizes to store the boot device configurations.  An unintended side effect is that a board with old config may ASSERT when booted with the new code.  To avoid this, users will be advised to erase all existing config before flashing new firmware.

Signed-off-by: Ryan Harkin <ryan.harkin@linaro.org>