MdeModulePkg//ArpDxe: Retrieved SnpMode only after configuring Snp
When Arp driver starts (with ArpDriverBindingStart()), its service will be created
and the Mnp child configured (ArpService->Mnp->Configure() called in ArpCreateService()).

It is only at this time the Snp protocol will be initialized (at the end of MnpStart()).
So, a valid SnpMode could not be expected prior to ArpService->Mnp->Configure().

Signed-off-by: Olivier Martin <olivier.martin@arm.com>