panda: Enable native building for Pandaboard.
The "chtool" utility is checked in in its compiled form (x86-64),
which breaks when trying to compile on any other architecture.
This patch deletes this utility, which will be automatically
built when required.

PandaBoardPkg/build.sh uses TARGET_TOOLS rather than TOOLCHAIN
to determine toolchain to use - and falls back to hardcoded
cross-toolchains when this fails. To minimize changes, copy
the value of TOOLCHAIN into TARGET_TOOLS if set.