imgburn.sh: add parameters and defaults
Add parameters and some sensible defaults to imgburn.sh.

The old imgburn.sh took 2 parameters but didn't allow the user to specify the image location.

Now, the simplest command form is:
	./imgburn --disk /dev/sdX

However, other options exist, eg:
	./imgburn.sh --disk /dev/sdX --soc 5250 --image filename.fd

The default SoC is 5250.
The default image is ARNDALE_EFI.fd, with the assumption that we are in uefi-next tree.

Signed-off-by: Ryan Harkin <ryan.harkin@linaro.org>